Gallery2: Fix picture is different when changes font size
when change the system font size, photopage destroy. when re-start it,
the photopage restore from stack. But the current photo index is not saved.
So photopage display the original photo not the current photo.

In onSaveState the current photo index is saved. In onCreate that current
index is restored if the photopage restore from stack.

Change-Id: I9a0e4ff1d13773a976c197e96fef570c173b3b8d
CRs-Fixed: 1004545